fix change that introduced interger division

+++ assignment-client/src/octree/OctreeServer.cpp
@@ -613,13 +613,13 @@ bool OctreeServer::handleHTTPRequest(HTTPConnection* connection, const QUrl& url
             .arg(locale.toString((uint)totalWastedBytes).rightJustified(COLUMN_WIDTH, ' '));
         statsString += QString().sprintf("            Total OctalCode Bytes: %s bytes (%5.2f%%)\r\n",
             locale.toString((uint)totalBytesOfOctalCodes).rightJustified(COLUMN_WIDTH, ' ').toLocal8Bit().constData(),
-            (double)((totalBytesOfOctalCodes / totalOutboundBytes) * AS_PERCENT));
+            (double)((totalBytesOfOctalCodes / (float)totalOutboundBytes) * AS_PERCENT));
         statsString += QString().sprintf("             Total BitMasks Bytes: %s bytes (%5.2f%%)\r\n",
             locale.toString((uint)totalBytesOfBitMasks).rightJustified(COLUMN_WIDTH, ' ').toLocal8Bit().constData(),
-            (double)((totalBytesOfBitMasks / totalOutboundBytes) * AS_PERCENT));
+            (double)((totalBytesOfBitMasks / (float)totalOutboundBytes) * AS_PERCENT));
         statsString += QString().sprintf("                Total Color Bytes: %s bytes (%5.2f%%)\r\n",
             locale.toString((uint)totalBytesOfColor).rightJustified(COLUMN_WIDTH, ' ').toLocal8Bit().constData(),
-            (double)((totalBytesOfColor / totalOutboundBytes) * AS_PERCENT));
+            (double)((totalBytesOfColor / (float)totalOutboundBytes) * AS_PERCENT));
 
         statsString += "\r\n";
         statsString += "\r\n";
@@ -731,7 +731,7 @@ bool OctreeServer::handleHTTPRequest(HTTPConnection* connection, const QUrl& url
             checkSum += OctreeElement::getChildrenCount(i);
             statsString += QString().sprintf("    Nodes with %d children:      %s nodes (%5.2f%%)\r\n", i,
                 locale.toString((uint)OctreeElement::getChildrenCount(i)).rightJustified(16, ' ').toLocal8Bit().constData(),
-                (double)((OctreeElement::getChildrenCount(i) / nodeCount) * AS_PERCENT));
+                (double)((OctreeElement::getChildrenCount(i) / (float)nodeCount) * AS_PERCENT));
         }
         statsString += "                                ----------------------\r\n";
         statsString += QString("                    Total:      %1 nodes\r\n")
